+++ src/Components/SignUp/SignUp.jsx
@@ -4,8 +4,8 @@ import cookie from "react-cookies";
 import { useMutation } from "@tanstack/react-query";
 import { PcPageWrap } from "../../Layouts/Main/style";
 import CancelCheckModal from "../Modal/CancelCheckModal";
-import SignUpUser from "./SignUpUser";
-import SignUpItems from "./SignUpItems";
+import SignUpUser from "./Form/SignUpUser";
+import SignUpItems from "./Form/SignUpItems";
 import { postUser } from "../../Api/user";
 import * as S from "./style";
 // import SignUpItemContext from "../../hooks/SignUpItemContext";
